+++ java/client/src/test/java/glide/api/models/ClusterTransactionTests.java
@@ -1,6 +1,8 @@
 /** Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0 */
 package glide.api.models;
 
+import static command_request.CommandRequestOuterClass.RequestType.PubSubSChannels;
+import static command_request.CommandRequestOuterClass.RequestType.PubSubSNumSub;
 import static command_request.CommandRequestOuterClass.RequestType.SPublish;
 import static command_request.CommandRequestOuterClass.RequestType.Sort;
 import static command_request.CommandRequestOuterClass.RequestType.SortReadOnly;
@@ -30,6 +32,15 @@ public class ClusterTransactionTests {
 
         transaction.publish("msg", "ch1", true);
         results.add(Pair.of(SPublish, buildArgs("ch1", "msg")));
+
+        transaction.pubsubShardChannels();
+        results.add(Pair.of(PubSubSChannels, buildArgs()));
+
+        transaction.pubsubShardChannels("test*");
+        results.add(Pair.of(PubSubSChannels, buildArgs("test*")));
+
+        transaction.pubsubShardNumSub(new String[] {"ch1", "ch2"});
+        results.add(Pair.of(PubSubSNumSub, buildArgs("ch1", "ch2")));
 
         transaction.sortReadOnly(
                 "key1",