+++ node/src/Errors.ts
@@ -3,7 +3,7 @@
  */
 
 /// Base class for errors.
-export abstract class RedisError extends Error {
+export abstract class ValkeyError extends Error {
     public message: string;
 
     constructor(message?: string) {
@@ -17,10 +17,10 @@ export abstract class RedisError extends Error {
 }
 
 /// Errors that report that the client has closed and is no longer usable.
-export class ClosingError extends RedisError {}
+export class ClosingError extends ValkeyError {}
 
 /// Errors that were reported during a request.
-export class RequestError extends RedisError {}
+export class RequestError extends ValkeyError {}
 
 /// Errors that are thrown when a request times out.
 export class TimeoutError extends RequestError {}