+++ node/src/GlideClusterClient.ts
@@ -8,11 +8,11 @@ import {
     BaseClientConfiguration,
     Decoder,
     DecoderOption,
-    GlideRecord,
+    GlideRecord, // eslint-disable-line @typescript-eslint/no-unused-vars
+    GlideReturnType,
     GlideString,
     PubSubMsg,
     ReadFrom, // eslint-disable-line @typescript-eslint/no-unused-vars
-    GlideReturnType,
     convertGlideRecordToRecord,
 } from "./BaseClient";
 import {
@@ -54,6 +54,9 @@ import {
     createPublish,
     createPubsubShardChannels,
     createRandomKey,
+    createScriptExists,
+    createScriptFlush,
+    createScriptKill,
     createSort,
     createSortReadOnly,
     createTime,
@@ -201,7 +204,7 @@ export interface SlotIdTypes {
      */
     type: "primarySlotId" | "replicaSlotId";
     /**
-     * Slot number. There are 16384 slots in a redis cluster, and each shard manages a slot range.
+     * Slot number. There are 16384 slots in a Valkey cluster, and each shard manages a slot range.
      * Unless the slot is known, it's better to route using `SlotKeyTypes`
      */
     id: number;
@@ -259,7 +262,7 @@ export type SingleNodeRoute =
     | RouteByAddress;
 
 /**
- * Client used for connection to cluster Redis servers.
+ * Client used for connection to cluster servers.
  *
  * @see For full documentation refer to {@link https://github.com/valkey-io/valkey-glide/wiki/NodeJS-wrapper#cluster|Valkey Glide Wiki}.
  */
@@ -731,7 +734,7 @@ export class GlideClusterClient extends BaseClient {
      * @example
      * ```typescript
      * const response = await client.lolwut({ version: 6, parameters: [40, 20] }, "allNodes");
-     * console.log(response); // Output: "Redis ver. 7.2.3" - Indicates the current server version.
+     * console.log(response); // Output: "Valkey ver. 7.2.3" - Indicates the current server version.
      * ```
      */
     public async lolwut(
@@ -1422,6 +1425,77 @@ export class GlideClusterClient extends BaseClient {
      */
     public async unwatch(options?: RouteOption): Promise<"OK"> {
         return this.createWritePromise(createUnWatch(), {
+            decoder: Decoder.String,
+            ...options,
+        });
+    }
+
+    /**
+     * Check existence of scripts in the script cache by their SHA1 digest.
+     *
+     * @see {@link https://valkey.io/commands/script-exists/|valkey.io} for more details.
+     *
+     * @param sha1s - List of SHA1 digests of the scripts to check.
+     * @param options - (Optional) See {@link RouteOption}.
+     * @returns A list of boolean values indicating the existence of each script.
+     *
+     * @example
+     * ```typescript
+     * console result = await client.scriptExists(["sha1_digest1", "sha1_digest2"]);
+     * console.log(result); // Output: [true, false]
+     * ```
+     */
+    public async scriptExists(
+        sha1s: GlideString[],
+        options?: RouteOption,
+    ): Promise<ClusterResponse<boolean[]>> {
+        return this.createWritePromise(createScriptExists(sha1s), options);
+    }
+
+    /**
+     * Flush the Lua scripts cache.
+     *
+     * @see {@link https://valkey.io/commands/script-flush/|valkey.io} for more details.
+     *
+     * @param options - (Optional) Additional parameters:
+     * - (Optional) `mode`: the flushing mode, could be either {@link FlushMode.SYNC} or {@link FlushMode.ASYNC}.
+     * - (Optional) `route`: see {@link RouteOption}.
+     * @returns A simple `"OK"` response.
+     *
+     * @example
+     * ```typescript
+     * console result = await client.scriptFlush(FlushMode.SYNC);
+     * console.log(result); // Output: "OK"
+     * ```
+     */
+    public async scriptFlush(
+        options?: {
+            mode?: FlushMode;
+        } & RouteOption,
+    ): Promise<"OK"> {
+        return this.createWritePromise(createScriptFlush(options?.mode), {
+            decoder: Decoder.String,
+            ...options,
+        });
+    }
+
+    /**
+     * Kill the currently executing Lua script, assuming no write operation was yet performed by the script.
+     *
+     * @see {@link https://valkey.io/commands/script-kill/|valkey.io} for more details.
+     * @remarks The command is routed to all nodes, and aggregates the response to a single array.
+     *
+     * @param options - (Optional) See {@link RouteOption}.
+     * @returns A simple `"OK"` response.
+     *
+     * @example
+     * ```typescript
+     * console result = await client.scriptKill();
+     * console.log(result); // Output: "OK"
+     * ```
+     */
+    public async scriptKill(options?: RouteOption): Promise<"OK"> {
+        return this.createWritePromise(createScriptKill(), {
             decoder: Decoder.String,
             ...options,
         });